@@ -10,7 +10,7 @@
     "babel-preset-react-app": "^3.0.2",
     "babel-runtime": "6.23.0",
     "case-sensitive-paths-webpack-plugin": "2.1.1",
-    "chalk": "1.1.3",
+    "chalk": "^1.1.3",
     "css-loader": "0.28.4",
     "dotenv": "4.0.0",
     "eslint": "4.4.1",
@@ -30,9 +30,9 @@
     "postcss-flexbugs-fixes": "3.2.0",
     "postcss-loader": "2.0.6",
     "promise": "8.0.1",
-    "react": "^15.6.1",
+    "react": "^16.12.0",
     "react-dev-utils": "^3.1.0",
-    "react-dom": "^15.6.1",
+    "react-dom": "^16.12.0",
     "react-error-overlay": "^1.0.10",
     "react-redux": "^5.0.6",
     "redux": "^3.7.2",
