@@ -9,9 +9,9 @@ export default class FormFieldToggle extends Component {
         };
     }
 
-    componentWillReceiveProps(nextProps) {
-        this.setState({
-            value: nextProps.value,
+    componentDidUpdate(prevProps) {
+        prevProps.value !== this.props.value && this.setState({
+            value: this.props.value,
         });
     }
 
